Fix bug in table_socketmap: error check of getline
Previously, if the return value of getline was not equal to -1,
table-socketmap would warn and go to an error state.  In fact, getline
signals an error by returning -1. Hence, the program would go into an
error state when getline indicated that no error had occurred.

Now, table-socketmap only moves to an error state when getline returns
an error.